requirements.txt
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.9.11
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.7
django-cors-headers==3.2.0
django-debug-toolbar==2.1
django-extensions==2.2.5
django-modeladmin-reorder==0.3.1
django-rest-passwordreset==1.1.0
django-summernote==0.8.11.6
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.0
et-xmlfile==1.0.1
firebase-admin==4.0.0
# freeze
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio # ==1.27.2
harupy==1.3
httplib2==0.17.0
hyperlink==20.0.1
idna==2.8
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
MarkupSafe==1.1.1
msgpack==1.0.0
openapi-codec==1.3.2
openpyxl==3.0.1
packaging==19.2
Pillow # ==6.2.1
protobuf==3.11.3
psycopg2-binary # ==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
short-url==1.2.2
simplejson==3.16.0
six==1.13.0
soupsieve==2.2.1
sqlparse==0.3.0
Twisted==20.3.0
uritemplate==3.0.0
urllib3==1.25.7
zope.interface==5.2.0
